Add helper script to update shared TF secret
between 0.2 and upcoming 0.3 beta releases TF acquired a need for
some additional info in the shared secret created by
openstack-controller, however unfortunately openstack-controller does
not add these new fields to the secret on controller update.

In order to facilitate update from 0.2 to 0.3 this additional info
can be added to the secret with the help of provided script.
It must be executed after openstack-controller and OpenStack update
but before TungstenFabric update.

Partial-Bug: PRODX-6070
Change-Id: I39a606906b766109286f176e56fe5740a56e59b1